requirements.txt
# ── Core AI ─────────────────
groq
python-dotenv

# ── UI ─────────────────────
rich

# ── Web/API Requests ──────
requests

# ── Reports ────────────────
reportlab

# ── Dashboard ──────────────
streamlit
pandas
pyvis

# ── REST API ───────────────
flask

# ── Credential Vault ──────
cryptography
